requirements.txt
build>=0.10.0
twine>=4.0.2
setuptools>=65.5.1



websockets>=11.0.3
requests>=2.31.0
websocket-client>=1.6.2